@@ -1,19 +1,23 @@
-module github.com/zchee/mcp-servers/sequential-thinking-server
+module github.com/zchee/mcp-servers/sequential-thinking
 
-go 1.24
+go 1.25
 
 require (
 	github.com/bytedance/gg v1.1.0
-	github.com/bytedance/sonic v1.14.0-rc1.0.20250708085712-f51fac7a464f // @main
-	github.com/modelcontextprotocol/go-sdk v0.2.1-0.20250714164758-de4b788e5995 // @main
+	github.com/bytedance/sonic v1.14.3-0.20251120112852-388df2d3cca6 // @main
+	github.com/google/jsonschema-go v0.3.1-0.20251120200837-98a387e3b975 // @main
+	github.com/google/uuid v1.6.0
+	github.com/modelcontextprotocol/go-sdk v1.1.1-0.20251119220150-194378025552 // @main
 )
 
 require (
-	github.com/bytedance/sonic/loader v0.3.0 // indirect
-	github.com/cloudwego/base64x v0.1.5 // indirect
-	github.com/klauspost/cpuid/v2 v2.3.0 // indirect
+	github.com/bytedance/gopkg v0.1.3 // indirect
+	github.com/bytedance/sonic/loader v0.4.0 // indirect
+	github.com/cloudwego/base64x v0.1.6 // indirect
+	github.com/klauspost/cpuid/v2 v2.2.9 // indirect
 	github.com/twitchyliquid64/golang-asm v0.15.1 // indirect
 	github.com/yosida95/uritemplate/v3 v3.0.2 // indirect
-	golang.org/x/arch v0.19.0 // indirect
-	golang.org/x/sys v0.34.0 // indirect
+	golang.org/x/arch v0.0.0-20210923205945-b76863e36670 // indirect
+	golang.org/x/oauth2 v0.30.0 // indirect
+	golang.org/x/sys v0.22.0 // indirect
 )
